Cleanup and at least get markdown on the screen
